+++ packages/reactivity/src/types/index.ts
@@ -1,3 +1,4 @@
 export * from './computed'
+export * from './extend'
 export * from './raw'
 export * from './ref'