+++ packages/reactivity/src/types/ref.ts
@@ -3,3 +3,9 @@ export type Ref<T = unknown> = {
 }
 
 export type MaybeRef<T = unknown> = T | Ref<T>
+
+export type RefFunction = {
+  <T extends Ref>(value: T): T
+  <T>(value: T): Ref<T>
+  <T>(): Ref<T | undefined>
+}